app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Simple example bot for Derebit"
authors = ["Eliseev Nikolay <n.a.eliseev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
anyio = "^2.1.0"
PyYAML = "^5.4.1"
json5 = "^0.9.5"
websockets = "^8.1"
mysql-connector-python = "^8.0.23"
deribitsimplebot = "^0.2.2"

[tool.poetry.scripts]
start-app = "app:main"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"